renamed version file to be same as rest
