+++ spalloc_server/__init__.py
@@ -1,1 +1,1 @@
-from spalloc_server.version import __version__  # noqa
+from spalloc_server._version import __version__  # noqa
